restructuring code
