+++ onnxmltools/convert/sparkml/__init__.py
@@ -6,3 +6,5 @@
 
 from .convert import convert
 from .utils import buildInitialTypesSimple, getTensorTypeFromSpark, buildInputDictSimple
+from .ops_names import get_sparkml_operator_name
+from .ops_input_output import get_input_names, get_output_names
